+++ packages/config/src/Finder.ts
@@ -7,7 +7,7 @@ import loadJson from './loaders/json';
 import loadMjs from './loaders/mjs';
 import loadYaml from './loaders/yaml';
 import Cache from './Cache';
-import getEnv from './getEnv';
+import getEnv from './helpers/getEnv';
 import { FinderOptions, ExtType, ConfigFile } from './types';
 import { CONFIG_FOLDER, DEFAULT_EXTS, PACKAGE_FILE } from './constants';
