+++ packages/config/src/helpers/mergeArray.ts
