+++ packages/config/src/helpers/readFile.ts
