+++ packages/config/src/loaders/yaml.ts
@@ -1,5 +1,5 @@
 import { Path, yaml } from '@boost/common';
-import readFile from '../readFile';
+import readFile from '../helpers/readFile';
 
 export default function loadYaml<T>(path: Path): Promise<T> {
   return readFile(path).then(data => yaml.parse(data));
